refactors

+++ src/ini.h
@@ -47,12 +47,12 @@ static inline void dict_reserve(Dict* dict, int cap) {
     if (cap <= dict->cap) {
         return;
     }
-    dict->items = (DictItem*)realloc(dict->items, (size_t)cap * sizeof(DictItem));
+    dict->items = (DictItem*)realloc(dict->items, cap * sizeof(DictItem));
     if (!dict->items) {
         perror("realloc");
         exit(1);
     }
-    memset(dict->items + dict->cap, 0, (size_t)(cap - dict->cap) * sizeof(DictItem));
+    memset(dict->items + dict->cap, 0, (cap - dict->cap) * sizeof(DictItem));
     dict->cap = cap;
 }
 
@@ -273,7 +273,7 @@ static int puf_ini_parse_list(const char* raw, double** out, int* len) {
         }
     }
 
-    double* values = (double*)calloc((size_t)cap, sizeof(double));
+    double* values = (double*)calloc(cap, sizeof(double));
     if (!values) {
         perror("calloc");
         exit(1);
@@ -328,7 +328,7 @@ static Dict* puf_ini_section(Ini* ini, const char* name, int add) {
     }
 
     ini->sections = (Dict*)realloc(ini->sections,
-        (size_t)(ini->num_sections + 1) * sizeof(Dict));
+        (ini->num_sections + 1) * sizeof(Dict));
     if (!ini->sections) {
         perror("realloc");
         exit(1);
